requirements.txt
moderngl~=5.4.2
pymunk~=5.4.0
PyQt5~=5.11.3
Django~=2.1.3
Pillow~=5.3.0
